+++ CHANGELOG.md
@@ -19,6 +19,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Added `TabActivated` message to `TabbedContent` https://github.com/Textualize/textual/pull/2260
 
 
+### Fixed
+
+- Fixed order styles are applied in DataTable - allows combining of renderable styles and component classes https://github.com/Textualize/textual/pull/2272
+
 ## [0.19.1] - 2023-04-10
 
 ### Fixed